# openapi.yaml
openapi: 3.0.0
info:
  title: SGTR-FEI API - Sistema de Gestión de Trabajos Recepcionales
  version: 1.0.0
  description: Documentación interactiva de la API RESTful para el Sistema de
    Gestión de Información de Trabajos Recepcionales de la Facultad de
    Estadística e Informática (FEI) de la Universidad Veracruzana.
  contact:
    name: Facultad de Estadística e Informática - UV
    url: https://www.uv.mx/fei/
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local
components:
  schemas:
...
    AsignarAcademicoDTO:
      type: object
      required:
        - Id_TrabajoR
        - Numero_Personal
        - Id_rol
      properties:
        Id_TrabajoR:
          type: integer
          example: 1
        Numero_Personal:
          type: integer
          example: 12345
        Id_rol:
          type: integer
          example: 1
          description: "1: Director, 2: Codirector, 3: Sinodal, etc."
    AsignarEstudianteDTO:
      type: object
      required:
        - Id_TrabajoR
        - Matricula
      properties:
        Id_TrabajoR:
          type: integer
          example: 1
        Matricula:
          type: string
          example: s20014589
...
paths:
...
  "/api/participantes/trabajo/{idTrabajo}":
    get:
      summary: Obtener el resumen completo de participantes (académicos y estudiantes)
        de un trabajo recepcional
      tags:
        - Participantes
      parameters:
        - in: path
          name: idTrabajo
          required: true
          schema:
            type: integer
          description: ID del trabajo recepcional
      responses:
        "200":
          description: Listado con arreglos de académicos y estudiantes asignados
        "500":
          description: Error en la consulta
  /api/participantes/academico:
    post:
      summary: Asignar un académico con rol (director, codirector, sinodal) a un
        trabajo recepcional
      tags:
        - Participantes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AsignarAcademicoDTO"
      responses:
        "201":
          description: Académico asignado exitosamente
        "400":
          description: Parámetros inválidos o asignación duplicada
  "/api/participantes/academico/trabajo/{idTrabajo}":
    get:
      summary: Listar únicamente los académicos participantes de un trabajo recepcional
      tags:
        - Participantes
      parameters:
        - in: path
          name: idTrabajo
          required: true
          schema:
            type: integer
          description: ID del trabajo recepcional
      responses:
        "200":
          description: Lista de académicos participantes
  "/api/participantes/academico/{id}":
    delete:
      summary: Remover la asignación de un académico de un trabajo recepcional
      tags:
        - Participantes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID de la participación (Id_Participacion)
      responses:
        "200":
          description: Académico removido exitosamente
        "400":
          description: Asignación no encontrada
  /api/participantes/estudiante:
    post:
      summary: Asignar un estudiante a un trabajo recepcional
      tags:
        - Participantes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AsignarEstudianteDTO"
      responses:
        "201":
          description: Estudiante asignado exitosamente
        "400":
          description: Parámetros inválidos o estudiante ya asignado al trabajo
  "/api/participantes/estudiante/trabajo/{idTrabajo}":
    get:
      summary: Listar los estudiantes asignados a un trabajo recepcional
      tags:
        - Participantes
      parameters:
        - in: path
          name: idTrabajo
          required: true
          schema:
            type: integer
          description: ID del trabajo recepcional
      responses:
        "200":
          description: Lista de estudiantes asignados
  "/api/participantes/estudiante/{id}":
    delete:
      summary: Remover la asignación de un estudiante de un trabajo recepcional
      tags:
        - Participantes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID de la asignación del estudiante (Id_EstudianteTrabajo)
      responses:
        "200":
          description: Estudiante desvinculado exitosamente
        "400":
          description: Asignación no encontrada
  /api/participantes/externo:
    post:
      summary: Asignar un participante externo a un trabajo recepcional
      tags:
        - Participantes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - Id_TrabajoR
                - Id_ParticipanteExt
                - Id_rol
              properties:
                Id_TrabajoR:
                  type: integer
                Id_ParticipanteExt:
                  type: integer
                Id_rol:
                  type: integer
      responses:
        "201":
          description: Participante externo asignado correctamente
  "/api/participantes/externo/trabajo/{idTrabajo}":
    get:
      summary: Listar los participantes externos de un trabajo recepcional
      tags:
        - Participantes
      parameters:
        - in: path
          name: idTrabajo
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de participantes externos asignados
  "/api/participantes/externo/{id}":
    delete:
      summary: Desvincular un participante externo de un trabajo recepcional
      tags:
        - Participantes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Participante externo removido exitosamente
  /api/participantes-externos:
    post:
      summary: Registrar un nuevo participante externo
      tags:
        - Participantes Externos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - Nombre
                - ApellidoP
              properties:
                Nombre:
                  type: string
                  example: Roberto
                ApellidoP:
                  type: string
                  example: Martínez
                ApellidoM:
                  type: string
                  example: Soto
                CorreoElectronico:
                  type: string
                  example: roberto.martinez@empresa.com
                Institucion:
                  type: string
                  example: Instituto Nacional de Investigaciones Eléctricas
      responses:
        "201":
          description: Participante externo registrado exitosamente
        "400":
          description: Error en los datos proporcionados
    get:
      summary: Listar participantes externos con filtros opcionales
      tags:
        - Participantes Externos
      parameters:
        - in: query
          name: Institucion
          schema:
            type: string
          description: Filtrar por institución de procedencia
        - in: query
          name: busqueda
          schema:
            type: string
          description: Búsqueda general por nombre, apellido, correo o institución
      responses:
        "200":
          description: Lista de participantes externos
  "/api/participantes-externos/{id}":
    get:
      summary: Obtener el detalle de un participante externo por ID
      tags:
        - Participantes Externos
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detalle del participante externo y trabajos vinculados
        "404":
          description: No encontrado
    put:
      summary: Actualizar la información de un participante externo
      tags:
        - Participantes Externos
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Participante externo actualizado correctamente
        "400":
          description: Error al actualizar
    delete:
      summary: Eliminar un participante externo
      tags:
        - Participantes Externos
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eliminado exitosamente
        "400":
          description: No se puede eliminar si tiene trabajos vinculados
  /api/participantes-externos/asignar:
    post:
      summary: Asignar un participante externo a un trabajo recepcional con un rol
      tags:
        - Participantes Externos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - Id_TrabajoR
                - Id_ParticipanteExt
                - Id_rol
              properties:
                Id_TrabajoR:
                  type: integer
                  example: 1
                Id_ParticipanteExt:
                  type: integer
                  example: 1
                Id_rol:
                  type: integer
                  example: 2
                  description: ID del rol (Director, Codirector, Sinodal, etc.)
      responses:
        "201":
          description: Participante externo asignado correctamente
  "/api/participantes-externos/trabajo/{idTrabajo}":
    get:
      summary: Listar los participantes externos asignados a un trabajo recepcional
      tags:
        - Participantes Externos
      parameters:
        - in: path
          name: idTrabajo
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de participantes externos asignados
  "/api/participantes-externos/asignacion/{id}":
    delete:
      summary: Desvincular a un participante externo de un trabajo recepcional
      tags:
        - Participantes Externos
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID de la asignación (Id_ParticipanteExtTrabajo)
      responses:
        "200":
          description: Participante externo desvinculado con éxito
...